Updated CommunityPosts Page
