+++ frontend/src/components/CommunityPostsCard.jsx
@@ -1,60 +1,27 @@
-import { Checkbox, CheckboxGroup, SimpleGrid, Box, Flex, FormLabel, FormControl, Input, Center, Stack, Button, Select, Card } from '@chakra-ui/react'
 import { useEffect, useState } from "react";
+import { NavLink } from "react-router-dom";
 import { Autocomplete } from "@react-google-maps/api";
+import { Box, Card, CardBody, CardHeader, CardFooter, Flex, FormControl, Heading, Image, Text, IconButton, Input, SkeletonText, InputGroup, InputRightElement, SimpleGrid } from "@chakra-ui/react";
 import PostCard from "./PostCard";
 import Map from "../components/Map";
 
-export default function CommunityPostsCard({ filterClick, filteredPosts, setStartTime, setEndTime, setDate, setPosts }) {
-    // const handleSubmit = (e) => {
-    //     e.preventDefault()
-    //     const place = document.getElementById('location').value
-    //     console.log(place)
-    //     setLocation(place);
-    // }
-
-    const handleDateChange = (e) => {
-        setDate(e.target.value);
-    }
-
-    const handleStartChange = (e) => {
-        setStartTime(e.target.value);
-    };
-
-    const handleEndChange = (e) => {
-        setEndTime(e.target.value);
-    };
+export default function CommunityPostsCard({ post, index }) {
 
     return (
         <>
-            <Box position="relative" zIndex="1" className='h-[100%] w-[100%] flex flex-col justify-center items-center'>
-                {/* className='h-full w-full bg-[#D9D9D9] flex flex-col justify-center items-center p-[4rem]' */}
-
-                    {
-                        filterClick === "time" && (
-                            <Center>
-                                <FormControl className="mt-[5rem]">
-                                    <Flex direction="row" gap="4">
-                                        <FormLabel htmlFor='startTime'>Start:</FormLabel>
-                                        <Input onChange={handleStartChange} type='time' id='startTime' name='startTime' />
-                                        <FormLabel htmlFor='endTime'>End:</FormLabel>
-                                        <Input onChange={handleEndChange} type='time' id='endTime' name='endTime' />
-                                    </Flex>
-                                </FormControl>
-                            </Center>
-                        )
-                    }
-                    {
-                        filterClick === "date" && (
-                            <Center>
-                                <FormControl className="mt-[5rem]">
-                                    <FormLabel htmlFor='date'>Date:</FormLabel>
-                                    <Input onChange={handleDateChange} type='date' id='date' name='date' />
-                                </FormControl>
-                            </Center>
-                        )
-                    }
-                    <Map posts={filteredPosts} setPosts={setPosts} />
-            </Box>
+            <Card key={index} direction={'row'} className="w-full" >
+                <CardHeader>
+                    <Image src={post.image} alt="post image" className="w-[10em] h-[6em]" />
+                    <Text className="mt-[1em] text-gray-500">Location: {post.location}</Text>
+                </CardHeader>
+                <CardBody className="text-gray-500 flex flex-col w-[100%]">
+                    <Heading size='md'><NavLink to={`/posts/${post.id}`}>{post.title}</NavLink></Heading>
+                    {/* <Text className="h-[60%]">{post.description}</Text> */}
+                    <Text className="w-[6em]">Date: {post.date_of_event}</Text>
+                    <Text className="w-[6em]">Start: {post.start_time}</Text>
+                    <Text className="w-[6em]">End: {post.end_time}</Text>
+                </CardBody>
+            </Card>
         </>
     );
 }
