+++ frontend/src/components/Map.jsx
@@ -2,7 +2,7 @@
 import { useEffect, useState } from "react";
 import { SearchIcon, RepeatClockIcon } from '@chakra-ui/icons'
 import { GoogleMap, Marker, Autocomplete, InfoWindow } from '@react-google-maps/api';
-import { Box, Card, CardBody, CardHeader, CardFooter, Flex, FormControl, Heading, Image, Text, IconButton, Input, SkeletonText, InputGroup, InputRightElement, SimpleGrid } from "@chakra-ui/react";
+import { Box, Card, CardBody, Center, CardHeader, CardFooter, Flex, FormControl, Heading, Image, Text, IconButton, Input, SkeletonText, InputGroup, InputRightElement, SimpleGrid, FormLabel } from "@chakra-ui/react";
 import { fromAddress } from "react-geocode";
 import { googleApi, geoCode } from "../googleApi";
 import { getAllPosts } from "../adapters/post-adapter";
@@ -10,6 +10,7 @@ import CreatePostForm from "./CreatePostForm";
 import { NavLink, useNavigate } from "react-router-dom";
 import PostCard from "./PostCard";
 import SearchPostAndFilterBar from "./SearchPostAndFilterBar";
+import CommunityPostsCard from "./CommunityPostsCard";
 
 const containerStyle = {
   width: '100%',
@@ -24,10 +25,11 @@ const center = {
 
 
 
-export default function Map({ posts, setPosts }) {
-  const [hovered, setHovered] = useState(false);
+export default function Map({ posts, setPosts, sortClick, setSortClick, filterClick, setFilterClick, setDate, setStartTime, setEndTime, setLocation }) {
   const [map, setMap] = useState(/** @type google.maps.Map */)
   const [marker, setMarker] = useState(/** @type google.maps.Marker */)
+  const [hovered, setHovered] = useState(false);
+  const [hoverHeight, setHoverHeight] = useState(0);
   const [zoom, setZoom] = useState(10)
   const { isLoaded } = googleApi()
   geoCode()
@@ -56,21 +58,17 @@ export default function Map({ posts, setPosts }) {
     }
   }
 
-  const handleSortClick = (event) => {
-    setSortClick(event.target.value);
+  const handleDateChange = (e) => {
+    setDate(e.target.value);
   }
 
-  const handleFilterClick = (event) => {
-    setFilterClick(event.target.value);
-  }
+  const handleStartChange = (e) => {
+    setStartTime(e.target.value);
+  };
 
-
-  const handleFilterSubmit = (e) => {
-    e.preventDefault()
-    const place = document.getElementById('location').value
-    console.log(place)
-    setLocation(place);
-  }
+  const handleEndChange = (e) => {
+    setEndTime(e.target.value);
+  };
 
   const reset = () => {
     map.panTo(center)
@@ -78,46 +76,42 @@ export default function Map({ posts, setPosts }) {
   }
 
   return <>
-    <Flex h='100vh' w='100%' alignItems='center' justifyContent='center' className="z-8">
+    <Flex flexDir={'row'} h='100vh' w='100%' alignItems='center' justifyContent='center' className="z-8">
       <Box w='35%' h='85%' background='grey' className="rounded-lg">
-        <Card className="h-[10rem] w-full bg-white hover:bg-gray-300 items-end" onMouseEnter={() => setHovered(true)} onMouseLeave={() => setHovered(false)}>
-          <Flex className="w-full row">
-            <CardHeader className="w-[65%]">
-              <SearchPostAndFilterBar />
-            </CardHeader>
-            <CardBody>
-              <CreatePostForm posts={posts} setPosts={setPosts} hovered={hovered} className="w-[25%]" />
-            </CardBody>
-          </Flex>
+        <Card className=" w-full bg-white hover:bg-gray-300" onMouseEnter={() => setHovered(true)} onMouseLeave={() => setHovered(false)}>
+          <SearchPostAndFilterBar posts={posts} setPosts={setPosts} setLocation={setLocation} setSortClick={setSortClick} setFilterClick={setFilterClick} setDate={setDate} hovered={hovered} setStartTime={setStartTime} setEndTime={setEndTime} />
+          <Box>
+            {
+              filterClick === "time" && (
+                <Center>
+                  <FormControl className="mt-[5rem]">
+                    <Flex direction="row" gap="4">
+                      <FormLabel htmlFor='startTime'>Start:</FormLabel>
+                      <Input onChange={handleStartChange} type='time' id='startTime' name='startTime' />
+                      <FormLabel htmlFor='endTime'>End:</FormLabel>
+                      <Input onChange={handleEndChange} type='time' id='endTime' name='endTime' />
+                    </Flex>
+                  </FormControl>
+                </Center>
+              )
+            }
+            {
+              filterClick === "date" && (
+                <Center>
+                  <FormControl className="mt-[5rem]">
+                    <FormLabel htmlFor='date'>Date:</FormLabel>
+                    <Input onChange={handleDateChange} type='date' id='date' name='date' />
+                  </FormControl>
+                </Center>
+              )
+            }
+          </Box>
         </Card>
-        <Box w='100%' h={'87%'} overflowY={"scroll"}>
+        <Box w='100%' h={'77%'} overflowY={"scroll"}>
           <Flex flexDir={'column'} className="p-[1rem]">
             {posts.map((post, index) => (
               <ul overflow="scroll" key={post.id}>
-                <Card key={index} direction={'row'} className="w-full" >
-                  <CardHeader>
-                    <Image src={post.image} alt="post image" className="w-[10em] h-[6em]" />
-                    <Text className="mt-[1em] text-gray-500">Location: {post.location}</Text>
-                  </CardHeader>
-                  <CardBody className="text-gray-500 flex flex-col w-[100%]">
-                    <Heading size='md'><NavLink to={`/posts/${post.id}`}>{post.title}</NavLink></Heading>
-                    {/* <Text className="h-[60%]">{post.description}</Text> */}
-                    <Text className="w-[6em]">Date: {post.date_of_event}</Text>
-                    <Text className="w-[6em]">Start: {post.start_time}</Text>
-                    <Text className="w-[6em]">End: {post.end_time}</Text>
-                  </CardBody>
-                  <CardFooter className="text-gray-500 flex flex-row">
-                    {/* {!!isCurrentUserProfile &&
-                      (
-                        <ButtonGroup spacing='2' colorScheme='green' className="bottom-0">
-                          <Button onClick={() => handleDelete(post.id)} variant='ghost' colorScheme='green'>
-                            Delete
-                          </Button>
-                        </ButtonGroup>
-                      )
-                    } */}
-                  </CardFooter>
-                </Card>
+                <CommunityPostsCard post={post} index={index} />
               </ul>
             ))}
           </Flex>
